requirements.txt
autopep8==1.7.0
nltk==3.7
Flask==2.2.2
regex==2022.8.17
urllib3==1.26.12
numpy==1.21.6
requests==2.28.2
pinecone-client==2.2.1
pandas==1.3.5
sentence-transformers==2.2.2
